Add forgotten uncommitted changes

+++ src/main/java/pl/edu/prz/weii/ChinesePostmanProblem/application/RouteFinder.java
@@ -1,6 +1,5 @@
 package pl.edu.prz.weii.ChinesePostmanProblem.application;
 
-import org.apache.commons.cli.*;
 import org.jenetics.*;
 import org.jenetics.engine.Engine;
 import org.jenetics.engine.EvolutionResult;
@@ -24,33 +23,44 @@ public class RouteFinder {
     private Set<Integer> nodes;
     private double maxWeight;
 
-    private int populationSize = 100;
-    private int startChromosomeCount = 10;
-    private int startChromosomeLength = 10;
-    private long limitIterations = 100_000_000_000L;
-    private int limitSteady = 100_000_000;
-    private double probabilityOfMutation = 0.1;
-    private int penaltyMultiplier = 100;
+    private int populationSize = 10000;
+    private int startChromosomeCount = 6;
+    private int startChromosomeLength = 2;
+    private long limitIterations = 1_000_000L;
+    private int limitSteady = 10_000;
+    private int penaltyMultiplier = 10;
+
+    private double routeMutatorProbability = 0.1;
+    private double singlePointCrossoverProbability = 0.05;
+    private boolean useRoulette = false;
     private double penalty;
     private String alterers = "";
     private String offspringSelector = "";
     private String survivorsSelector = "";
-
-
     private double prevBestScore = Double.MAX_VALUE;
     private long startTime = -1;
 
-    public RouteFinder(FileContent fileContent) {
+
+    public RouteFinder(FileContent fileContent, int populationSize, int startChromosomeCount, int startChromosomeLength, int penaltyMultiplier, double routeMutatorProbability, double singlePointCrossoverProbability, boolean useRoulette) {
+        this.populationSize = populationSize;
+        this.startChromosomeCount = startChromosomeCount;
+        this.startChromosomeLength = startChromosomeLength;
+        this.penaltyMultiplier = penaltyMultiplier;
+        this.routeMutatorProbability = routeMutatorProbability;
+        this.singlePointCrossoverProbability = singlePointCrossoverProbability;
+        this.useRoulette = useRoulette;
         this.edges = fileContent.getEdges();
         this.nodes = fileContent.getNodes();
 
         try {
             new File("results").mkdirs();
-            this.outFile = new File("results/nodes-" + this.nodes.size() + "-edges-" + this.edges.size() + "-time-" + System.currentTimeMillis() + ".csv");
-            this.outFile.createNewFile();
-            List<String> lines = Collections.singletonList("nodes;edges;time[s];weight;score;penalty;populationSize;startChromosomeCount;startChromosomeLength;"
-                    + "limitIterations;limitSteady;alterers;offspringSelector;survivorsSelector;routeNodes;routeEdges");
-            Files.write(this.outFile.toPath(), lines, StandardCharsets.UTF_8, StandardOpenOption.APPEND);
+            this.outFile = new File("results/nodes-" + this.nodes.size() + "-edges-" + this.edges.size() + ".csv");
+            if (!this.outFile.exists()) {
+                this.outFile.createNewFile();
+                List<String> lines = Collections.singletonList("nodes;edges;valid;time[s];weight;penalty;populationSize;startChromosomeCount;startChromosomeLength;"
+                        + "limitIterations;limitSteady;alterers;offspringSelector;survivorsSelector;routeNodes;routeEdges");
+                Files.write(this.outFile.toPath(), lines, StandardCharsets.UTF_8, StandardOpenOption.APPEND);
+            }
         } catch (IOException e) {
             e.printStackTrace();
         }
@@ -64,8 +74,8 @@ public class RouteFinder {
         this.penalty = maxWeight * this.penaltyMultiplier;
     }
 
+
     public double fitness(final Route route) {
-        long time = System.currentTimeMillis() - startTime;
         double score = route.getWeight();
         if (!route.isValid()) {
             score += (route.getNotVisitedEdges() * penalty);
@@ -73,61 +83,52 @@ public class RouteFinder {
             if (!route.isStartingAndEndingOnSameNode()) {
                 score += penalty * 2;
             }
-        } else {
-            if (prevBestScore > score) {
-                prevBestScore = score;
-                StringJoiner stringJoiner = new StringJoiner(";");
-                stringJoiner
-                        .add(String.valueOf(this.nodes.size()))
-                        .add(String.valueOf(this.edges.size()))
-                        .add(String.valueOf(time/ 1000.0))
-                        .add(String.valueOf(route.getWeight()))
-                        .add(String.valueOf(score))
-                        .add(String.valueOf(penalty))
-                        .add(String.valueOf(populationSize))
-                        .add(String.valueOf(startChromosomeCount))
-                        .add(String.valueOf(startChromosomeLength))
-                        .add(String.valueOf(limitIterations))
-                        .add(String.valueOf(limitSteady))
-                        .add(String.valueOf(alterers))
-                        .add(String.valueOf(offspringSelector))
-                        .add(String.valueOf(survivorsSelector))
-                        .add(String.valueOf(route.getVisitedNodes()))
-                        .add(String.valueOf(route.getCorrectlyVisitedEdges()));
-
-                try {
-                    List<String> lines = Collections.singletonList(stringJoiner.toString());
-                    System.out.println(lines);
-                    Files.write(this.outFile.toPath(), lines, StandardCharsets.UTF_8, StandardOpenOption.APPEND);
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
-
-        }
-
-        if (prevBestScore > score) {
-            prevBestScore = score;
-            System.out.println(time + " " + score + " " + route);
         }
         return score;
     }
 
+    private void toFile(Route route){
+        long time = System.currentTimeMillis() - startTime;
+        StringJoiner stringJoiner = new StringJoiner(";");
+        stringJoiner
+                .add(String.valueOf(this.nodes.size()))
+                .add(String.valueOf(this.edges.size()))
+                .add(String.valueOf(route.isValid()))
+                .add(String.valueOf(time / 1000.0))
+                .add(String.valueOf(route.getWeight()))
+                .add(String.valueOf(penalty))
+                .add(String.valueOf(populationSize))
+                .add(String.valueOf(startChromosomeCount))
+                .add(String.valueOf(startChromosomeLength))
+                .add(String.valueOf(limitIterations))
+                .add(String.valueOf(limitSteady))
+                .add(String.valueOf(alterers))
+                .add(String.valueOf(offspringSelector))
+                .add(String.valueOf(survivorsSelector))
+                .add(String.valueOf(route.getVisitedNodes()))
+                .add(String.valueOf(route.getCorrectlyVisitedEdges()));
+
+        try {
+            List<String> lines = Collections.singletonList(stringJoiner.toString());
+            Files.write(this.outFile.toPath(), lines, StandardCharsets.UTF_8, StandardOpenOption.APPEND);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
     public Route findBest() {
-        startTime = System.currentTimeMillis();
+        this.startTime = System.currentTimeMillis();
 
-
-        final Engine<IntegerGene, Double> engine = Engine
+        Engine.Builder<IntegerGene, Double> engineBuilder = Engine
                 .builder(this::fitness, Route.code(this.nodes, this.edges, startChromosomeCount, startChromosomeLength))
                 .minimizing()
                 .alterers(
-                        new RouteMutator<>(this.probabilityOfMutation, this.nodes.size())
-                        , new SwapMutator<>()
-                        , new SinglePointCrossover<>()
+                        new RouteMutator<>(this.routeMutatorProbability, this.nodes.size()),
+                        new SinglePointCrossover<>(this.singlePointCrossoverProbability)
                 )
-                .populationSize(this.populationSize)
-                .selector(new RouletteWheelSelector<>())
-                .build();
+                .populationSize(this.populationSize).selector(new RouletteWheelSelector<>());
+
+        Engine<IntegerGene, Double> engine = engineBuilder.build();
 
         this.alterers = engine.getAlterer().toString().replace("\n", "");
         this.offspringSelector = engine.getOffspringSelector().toString();
@@ -139,8 +140,9 @@ public class RouteFinder {
                 .collect(EvolutionResult.toBestEvolutionResult());
         Genotype<IntegerGene> genotype = result.getBestPhenotype().getGenotype();
 
-
-        return Route.code(this.nodes, this.edges, startChromosomeCount, startChromosomeLength).decoder().apply(genotype);
+        Route bestRoute = Route.code(this.nodes, this.edges, startChromosomeCount, startChromosomeLength).decoder().apply(genotype);
+        toFile(bestRoute);
+        return bestRoute;
     }
 
 
@@ -155,27 +157,4 @@ public class RouteFinder {
                 '}';
     }
 
-
-    public static void main(String[] args) {
-        Options options = new Options();
-
-        Option input = new Option("i", "input", true, "input file path");
-        input.setRequired(true);
-        options.addOption(input);
-        CommandLineParser parser = new DefaultParser();
-        HelpFormatter formatter = new HelpFormatter();
-        CommandLine cmd;
-        try {
-            cmd = parser.parse(options, args);
-        } catch (ParseException e) {
-            System.out.println(e.getMessage());
-            formatter.printHelp("utility-name", options);
-            System.exit(1);
-            return;
-        }
-        String inputFilePath = cmd.getOptionValue("input");
-        String outputFilePath = cmd.getOptionValue("output");
-        System.out.println(inputFilePath);
-        System.out.println(outputFilePath);
-    }
 }
